Prompt Privacy, and clean-up
Updated the prompt to refer to delimiters within values (versus in the prompt), but MORE IMPORTANTLY added support for the ability to focus on particular context within the original article (which will be pulled from the .config file for privacy). This is called prompt_focus, and it could be things like special attention to strategy, geopolitics, anything. The only negative is that you have to update your prompt within another key value within the .config file. By default, there is no specifier

Also refactored the name of load_api_key() to load_key_from_config_file, given the generic nature of the .config file